split UI into components/ module
