+++ x-pack/plugins/endpoint/public/components/nav.tsx
@@ -1,0 +1,49 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import React from 'react';
+import { BrowserRouter as Router, Route, withRouter, RouteComponentProps } from 'react-router-dom';
+
+import {
+  EuiPage,
+  EuiPageBody,
+  EuiPageContent,
+  EuiPageContentBody,
+  EuiPageContentHeader,
+  EuiPageContentHeaderSection,
+  EuiPageHeader,
+  EuiPageHeaderSection,
+  EuiPageSideBar,
+  EuiTitle,
+  EuiSideNav,
+} from '@elastic/eui';
+
+type NavProps = RouteComponentProps & {
+  navigateToApp: AppMountContext['core']['application']['navigateToApp'];
+};
+
+export const Nav = withRouter(({ history, navigateToApp }: NavProps) => (
+  <EuiSideNav
+    items={[
+      {
+        name: 'Endpoint',
+        id: 'endpoint',
+        items: [
+          {
+            id: 'home',
+            name: 'Home',
+            onClick: () => history.push('/'),
+          },
+          {
+            id: 'management',
+            name: 'Management',
+            onClick: () => history.push('/management'),
+          },
+        ],
+      },
+    ]}
+  />
+));
